+++ src/modules/mod1/D.cxx
@@ -3,4 +3,7 @@ export module D;
 
 export void greetD();
 
-export float lol_d() { return 32.f / 4.5; }
+export double lol_d()
+{
+	return 32.f / 4.5;
+}
